requirements.txt
langchain==0.2.16
langchain-openai==0.1.14
langchain-community==0.2.16
langgraph==0.2.0
chromadb==0.5.0
beautifulsoup4==4.12.0
lxml==5.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
graphviz>=0.20.3,<0.21
